+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Barrier.java
@@ -37,16 +37,16 @@ public class Barrier extends ShieldBuff {
 		type = buffType.POSITIVE;
 	}
 
-	float partialLostShield;
+	double partialLostShield;
 
 	@Override
-	public void incShield(int amt) {
+	public void incShield(long amt) {
 		super.incShield(amt);
 		partialLostShield = 0;
 	}
 
 	@Override
-	public void setShield(int shield) {
+	public void setShield(long shield) {
 		super.setShield(shield);
 		if (shielding() == shield) partialLostShield = 0;
 	}
@@ -91,7 +91,7 @@ public class Barrier extends ShieldBuff {
 
 	@Override
 	public String iconTextDisplay() {
-		return Integer.toString(shielding());
+		return Long.toString(shielding());
 	}
 	
 	@Override
@@ -110,6 +110,6 @@ public class Barrier extends ShieldBuff {
 	@Override
 	public void restoreFromBundle(Bundle bundle) {
 		super.restoreFromBundle(bundle);
-		partialLostShield = bundle.getFloat(PARTIAL_LOST_SHIELD);
+		partialLostShield = bundle.getDouble(PARTIAL_LOST_SHIELD);
 	}
 }
