+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Corruption.java
@@ -36,12 +36,12 @@ public class Corruption extends AllyBuff {
 		announced = true;
 	}
 
-	private float buildToDamage = 0f;
+	private double buildToDamage = 0f;
 
 	//corrupted enemies are usually fully healed and cleansed of most debuffs
 	public static void corruptionHeal(Char target){
 		target.HP = target.HT;
-		target.sprite.showStatusWithIcon(CharSprite.POSITIVE, Integer.toString(target.HT), FloatingText.HEALING);
+		target.sprite.showStatusWithIcon(CharSprite.POSITIVE, Long.toString(target.HT), FloatingText.HEALING);
 		for (Buff buff : target.buffs()) {
 			if (buff.type == Buff.buffType.NEGATIVE
 					&& !(buff instanceof SoulMark)) {
@@ -54,7 +54,7 @@ public class Corruption extends AllyBuff {
 	public boolean act() {
 		buildToDamage += target.HT/100f;
 
-		int damage = (int)buildToDamage;
+		long damage = (long) buildToDamage;
 		buildToDamage -= damage;
 
 		if (damage > 0)
